Fridge Status Control/requirements.txt
paho_mqtt==2.1.0
pytz==2024.1
Requests==2.32.3
